+++ apps/frontend/src/app/events/page.tsx
@@ -3,9 +3,9 @@ import { redirect } from 'next/navigation';
 /**
  * Events Default Page
  *
- * Redirects to Event Types page as the default
+ * Redirects to Event Types settings page as the default
  */
 export default function EventsPage() {
-    redirect('/events/event-types');
+    redirect('/settings/event/type');
 }
